+++ packages/app/web/app/layout.tsx
@@ -3,6 +3,8 @@ import { ThemeProvider } from '@/components/theme-provider';
 import { Toaster } from '@/components/ui/sonner';
 import './global.css';
 import { Metadata } from 'next';
+import { Suspense } from 'react';
+import { Skeleton } from '@/components/ui/skeleton';
 
 const DmSans = DM_Sans({
   display: 'swap',
@@ -18,7 +20,18 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
     <html lang="zh" className={DmSans.className} suppressHydrationWarning>
       <body className="bg-background text-foreground">
         <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
-          {children}
+          <Suspense
+            fallback={
+              <div className="flex w-screen h-screen justify-center items-center flex-col gap-4">
+                <Skeleton className="h-4 w-1/2" />
+                <Skeleton className="h-4 w-1/2" />
+                <Skeleton className="h-4 w-1/2" />
+                <Skeleton className="h-4 w-3/8" />
+              </div>
+            }
+          >
+            {children}
+          </Suspense>
           <Toaster />
         </ThemeProvider>
       </body>